webapp/Biling/requirements.txt
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
attrs==22.1.0
autopep8==2.0.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
Brotli==1.0.9
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cssselect2==0.7.0
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
fastjsonschema==2.16.2
Flask==2.2.3
fonttools==4.38.0
html5lib==1.1
idna==3.4
importlib-metadata==5.1.0
importlib-resources==5.10.0
ipykernel==6.16.2
ipython==7.34.0
ipython-genutils==0.2.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.3
jupyter-server==1.23.3
jupyter_client==7.4.7
jupyter_core==4.12.0
jupyterlab-pygments==0.2.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==2.0.4
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.5
nbformat==5.7.0
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
prometheus-client==0.15.0
prompt-toolkit==3.0.33
psutil==5.9.4
pycodestyle==2.10.0
pycparser==2.21
pydyf==0.6.0
Pygments==2.13.0
pyodbc==4.0.39
pyparsing==3.0.9
pyphen==0.14.0
pyrsistent==0.19.2
python-dateutil==2.8.2
pywin32==305
pywinpty==2.0.9
pyzmq==24.0.1
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
terminado==0.17.0
tinycss2==1.2.1
tomli==2.0.1
tornado==6.2
traitlets==5.6.0
typing_extensions==4.4.0
wcwidth==0.2.5
weasyprint==59.0
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.3
zipp==3.11.0
zopfli==0.2.2
